components/toxic-data-cleaner/requirements.txt
scikit-learn
google-cloud-storage
pandas
flask
pickle
numpy